ENH Run in OS X or Linux
Modify `extendedcobey_runScenarios.py` so that it functions on OS X or Linux, via either a `wine` installation or via Docker. Include a Dockerfile which will build an image that can be used to do this.

Also modify pathing to provide the option of specifying local paths via environment variables (which may be defined in a `.env` file and loaded via `python-dotenv`). Modify function calls to explicitly pass needed directory names.